+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # In progress
 - Update processData getters 
+- Removing GW methods: `getProcListResults` and `getProcListLiveResults`
+- Renaming GW methods: `getScrutinizerEntities` into `getEntityList`
 
 # 0.20.0
 - Update `getGatewayInfo` to `getInfo`